feat: Disable 3D backgrounds by default
The 3D background feature in the Knowledge Tester application is now disabled by default. This change aims to improve initial load times and reduce resource consumption for new users. The setting can still be enabled by the user in the settings menu.

The logic for checking the 'enable-3d' setting has been updated to correctly reflect the default state.